requirements.txt
altair==4.2.2
attrs==22.2.0
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
easyocr==1.6.2
entrypoints==0.4
fonttools==4.39.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
imageio==2.26.0
importlib-metadata==6.0.0
imutils==0.5.4
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
lazy_loader==0.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
networkx==3.0
ninja==1.11.1
numpy==1.24.2
opencv-python==4.7.0.72
opencv-python-headless==4.5.4.60
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==3.20.3
pyarrow==11.0.0
pyclipper==1.3.0.post4
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
rich==13.3.2
scikit-image==0.20.0
scipy==1.10.1
semver==2.13.0
shapely==2.0.1
six==1.16.0
smmap==5.0.0
streamlit==1.20.0
tifffile==2023.2.28
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.15
validators==0.20.0
watchdog==2.3.1
zipp==3.15.0
altair==4.2.2
attrs==22.2.0
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
easyocr==1.6.2
entrypoints==0.4
fonttools==4.39.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
idna==3.4
imageio==2.26.0
importlib-metadata==6.0.0
imutils==0.5.4
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
lazy_loader==0.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
networkx==3.0
ninja==1.11.1
numpy==1.24.2
opencv-python==4.7.0.72
opencv-python-headless==4.5.4.60
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.1
protobuf==3.20.3
pyarrow==11.0.0
pyclipper==1.3.0.post4
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.19.3
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
rich==13.3.2
scikit-image==0.20.0
scipy==1.10.1
semver==2.13.0
shapely==2.0.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.6
streamlit==1.20.0
tenacity==8.2.2
tifffile==2023.2.28
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.15
validators==0.20.0
watchdog==2.3.1
zipp==3.15.0
